Fix float arguments being truncated to whole numbers over ARMI

+++ common/src/main/java/com/oddlabs/net/DefaultARMIArgumentWriter.java
@@ -26,7 +26,7 @@ public class DefaultARMIArgumentWriter implements ARMIArgumentWriter {
             long l = ((Long) arg);
             out.buffer().putLong(l);
         } else if (type.equals(float.class)) {
-            float f = ((Number) arg).intValue();
+            float f = ((Number) arg).floatValue();
             out.buffer().putFloat(f);
         } else if (type.equals(double.class)) {
             double d = ((Double) arg);
